+++ pysages/ssages/methods/core.py
@@ -1,0 +1,71 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (c) 2020: SSAGES Team (see LICENSE.md)
+
+from abc import ABC, abstractmethod
+
+from jax import jit
+from pysages.ssages.cvs import build
+
+
+# ================ #
+#   Base Classes   #
+# ================ #
+class SamplingMethod(ABC):
+    def __init__(self, cvs, *args, **kwargs):
+        self.cv = build(*cvs)
+        self.args = args
+        self.kwargs = kwargs
+
+    @abstractmethod
+    def __call__(self, *args, **kwargs):
+        pass
+
+
+class GriddedSamplingMethod(SamplingMethod):
+    def __init__(self, cvs, grid, *args, **kwargs):
+        check_dims(cvs, grid)
+        self.cv = build(*cvs)
+        self.grid = grid
+        self.args = args
+        self.kwargs = kwargs
+
+    @abstractmethod
+    def __call__(self, *args, **kwargs):
+        pass
+
+
+class NNSamplingMethod(SamplingMethod):
+    def __init__(self, cvs, grid, topology, *args, **kwargs):
+        check_dims(cvs, grid)
+        self.cv = build(*cvs)
+        self.grid = grid
+        self.topology = topology
+        self.args = args
+        self.kwargs = kwargs
+
+    @abstractmethod
+    def __call__(self, *args, **kwargs):
+        pass
+
+
+# ========= #
+#   Utils   #
+# ========= #
+
+def check_dims(cvs, grid):
+    if len(cvs) != grid.shape.size:
+        raise ValueError("Grid and Collective Variable dimensions must match.")
+
+
+def generalize(concrete_update):
+    _update = jit(concrete_update)
+
+    def update(snapshot, state):
+        vms = snapshot.vel_mass
+        rs = snapshot.positions
+        ids = snapshot.ids
+        #
+        return _update(state, rs, vms, ids)
+
+    # TODO: Validate that jitting here gives correct results
+    return jit(update)
